refactor: move localFlag + sourceRecents out of views/source/ → utils/
Both are pure localStorage helpers; no DOM. views/source/ shouldn't
hold non-DOM code. localFlag was already used by views/shell/
leftSidebar.ts — the source/ location was misleading from the start.
